# docs/open-api-docs.yaml
openapi: 3.0.3
info:
  title: The Agent's user-facing API
  description: The user-facing parts of The Agent's API service (excluding system-level endpoints, chat completion, maintenance endpoints, etc.)
  version: 5.28.1
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your.agent.com
    description: Your Agent API
security:
  - bearerAuth: []
paths:
  /attachments/private/{id}:
    get:
      summary: Read attachment privately
      description: Stream attachment bytes for a user who belongs to the attachment's chat
      operationId: getPrivateAttachment
      tags: [Attachments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Attachment ID
      responses:
        "200":
          description: Attachment bytes streamed successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /attachments/public/{token}:
    get:
      summary: Read attachment with public token
      description: Stream attachment bytes using a short-lived public attachment token
      operationId: getPublicAttachment
      tags: [Attachments]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Short-lived public attachment token
      responses:
        "200":
          description: Attachment bytes streamed successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security: []
  /settings/user/{user_id}:
    get:
      summary: Get user settings
      description: Retrieve settings for a specific user
      operationId: getUserSettings
      tags: [Settings]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettingsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
    patch:
      summary: Update user settings
      description: Save user settings including API keys and tool choices
      operationId: saveUserSettings
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsPayload"
      responses:
        "200":
          description: User settings saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/user/{user_id}/tools:
    get:
      summary: Get user's external tools configuration
      description: Retrieve external tools and providers configuration status
      operationId: getTools
      tags: [Settings]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Tools configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalToolsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/user/{user_id}/products:
    get:
      summary: Get available products
      description: Retrieve the list of purchasable products
      operationId: getProducts
      tags: [Settings]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/chats:
    get:
      summary: Get all chats for the invoker
      description: Retrieve all private and group chats where the invoker has a membership record; admin chats are discovered and their membership is refreshed
      operationId: getChats
      tags: [Chats]
      responses:
        "200":
          description: Chats retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSettingsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/chats/{chat_id}:
    get:
      summary: Get chat settings
      description: Retrieve settings for a specific chat; creates a membership row with false defaults if the invoker has no existing row
      operationId: getChatSettings
      tags: [Chats]
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: Chat ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Chat settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSettingsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
    patch:
      summary: Update chat settings
      description: Save chat and/or per-user chat configuration; admin-only fields require admin membership; at least one of chat_config or user_chat_config must be present
      operationId: saveChatSettings
      tags: [Chats]
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: Chat ID in hexadecimal format (UUID without hyphens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSettingsPayload"
      responses:
        "200":
          description: Chat settings saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /user/{user_id}/sponsorships:
    get:
      summary: Get user's sponsorships
      description: Retrieve all sponsorships created by the user
      operationId: getSponsorships
      tags: [Sponsorships]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Sponsorships retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorshipsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []

    post:
      summary: Create a new sponsorship
      description: Sponsor another user by their platform-specific handle
      operationId: sponsorUser
      tags: [Sponsorships]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorshipPayload"
      responses:
        "200":
          description: User sponsored successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorshipCreateResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
...
  /user/{user_id}/sponsored:
    delete:
      summary: Remove self-sponsorship
      description: User removes their own sponsorship status
      operationId: unsponsorSelf
      tags: [Sponsorships]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Self-sponsorship removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from one of the auth flows (usually via the chat)
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          nullable: false
          enum: [OK]
          description: Operation status
      required:
        - status
...
  responses:
    UnauthorizedError:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                nullable: false
                description: Error reason
            required:
              - detail
    ForbiddenError:
      description: Access denied — authorization failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"
    NotFoundError:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"